Fix small bugs in the relation charts
The link filter threw an error when the topk was undefined or zero. Now simply hides all links when those happen.
The negative similarity block in the second relation chart was not showing due to the height not being normalized to be positive.
The stroke opacity and width of the similarity links in the second relation charts were not normalized correctly.